pyproject.toml
[tool.poetry]
name = "compaigns"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
django-q = "^1.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
